feature(products): add keywords for seo
