+++ src/stores/auth/AutoLogin.tsx
@@ -14,8 +14,6 @@ export default function AutoLogin({}: Props) {
     const auth = useAuth()
     const { loadProduct } = useCart()
 
-    // console.log('auth: ', auth);
-
     useEffect(() => {
         const accessToken = clientInstance.getAccessToken()
         if (!accessToken) {
